Improve detailed calibration pop-up
