+++ src/basestation/CMakeLists.txt
@@ -1,10 +1,12 @@
 set(MODULE_HDR
 	modules/console.hpp
 	modules/controller_config.hpp
+	modules/controller_config_popup.hpp
 )
 set(MODULE_SRC
 	modules/console.cpp
 	modules/controller_config.cpp
+	modules/controller_config_popup.cpp
 )
 
 set(WIDGET_HDR
